+++ tests/unit/test_io_hdf5_h5tools.py
@@ -19,7 +19,7 @@ from hdmf.container import Container
 from h5py import SoftLink, HardLink, ExternalLink, File
 from h5py import filters as h5py_filters
 
-from tests.unit.test_utils import Foo, FooBucket, CORE_NAMESPACE
+from tests.unit.utils import Foo, FooBucket, CORE_NAMESPACE
 
 
 class FooFile(Container):
